move name methods to top of the files move inline fragment type to its own file
